requirements.txt
requests>=2.31.0
python-telegram-bot>=21.0
apscheduler>=3.10.0
python-dotenv>=1.0.0


